Implement email display routes
We have three routes:
- inbox: emails where the current user is on the to line
- sent: emails where the current user is on the from line
- outbox: sent emails that haven't been synced to the server yet

The sent/outgoing distinction is implemented using the Email.date field:
a null-value indicates that the email yet has to be synced to the server.

The current implementation uses ScalarListType to store array values on
the Email model (e.g. the Email.to field). This is inefficient because
we have to fetch all emails before we can restrict the result set to the
emails that the current user can view instead of doing it with a join in
the database. We may have to revisit this choice in the future.