Add dynamic factory
The DynamicFactory lets us construct objects by their fully qualified
class name. This is useful as it enables us to easily specify in a
configuration file which implementations of our interfaces we want to
use in the configured code.

The factory also has a pass-through mode if it's called with a class
reference instead of a fully qualified class name so that we can easily
override our configuration in tests.