Use real gzip in CompressedJson serializer
Using the zlib compressor, we can't operate on the compressed data using
tools that expect normal gzip data (e.g. zgrep or zless) because zlib
doesn't include the gzip header in the serialized files. This makes it a
bit difficult to deal with the serializer output outside of the context
of the flask app.