Add sync route and unit tests
We do three actions to sync the local database with the cloud:
1) Upload the outbox emails for all users.
2) Download the inbox emails for all users.
3) Download the updated account information for all users.

Actions 1) and 2) are self-explanatory. Action 3) is included so that
users can create new accounts locally and start sending emails straight
away under their local account name. When the email server sees a mail
sent by a non-email address, it creates a new mailbox for that user,
modifies the email's from field and sends it. The server also updates a
field in the file that the portable will download the next time around
to communicate the newly created email address back down to the
portable.

Both upload and download sync actions follow three steps that are
individually configurable:
1) The packer converts domain objects to serializable type.
2) The serializer converts the packed objects to bytes.
3) The remote storage transfers the bytes to or from the cloud.

Currently we have the following implementations:
1) Pack to dictionaries.
2) Serialize to compressed json.
3) Remote storage on Azure blob.

The change also includes a factory that lets us specify which
implementations we want to use for the various parts of the sync stack
by class name and construct them dynamically.