Move all translation strings out of templates
In this way, we only have one place where we need to look in order to
find all the strings that the application uses (`config.ui`) instead of
having to go through all templates and lots of places in the code.

Exception: long-form texts like mission statement, about page, etc. are
kept in the templates as they represent view-content as opposed to
potentially reusable UI strings.