Remove unnecessary grounding of iterator
Given that we only iterate the emails to display them later, we might as
well keep the original query object around as this lets us do things
like adding efficient pagination using limit and offset instead of
having to work with the full results-set in memory.

Also: encode return type of email lookup methods as iterable-of-emails
to make it explicit that we're getting back some sort of stream.