Implement pagination for email views
The heavy-lifting for the implementation is handled by the Pagination
query-wrapper provided by flask_sqlalchemy.
We can't directly use the class for all views as the inbox controller
returns a list instead of a query object so this commit also includes a
simple adapter that lets us use the Pagination class with list inputs.