Refactor to handle attachments in serializer
In order to support emails with attachments, we need to upload the
outbound attachments to Azure and download the inbound attachments. In
order to minimize bandwidth, we implement this by having all email data
and attachments in zip files which we exchange with Azure.

Note: This change makes the existing tests pass but is still largely
unverified.