Allow language switching via URL
We can now switch languages by changing the URL, e.g. we'll see the
English translation if we go to /en/home and the Italian page if we go
to /it/home. Once we're on a particular language, the value is sticky
(so following any link from /it/home will lead to another Italian page).
If the language isn't specified in the URL, we fall back to the
accept-languages sent by the browser.

The chosen language is persisted in the session so that we can remember
it across requests (e.g., to coordinate the localization of the login
and logout pages on the security blueprint).