Inline calls to gettext where possible
The evaluation of Jinja filters is cached which means that a construct
like `{{ 'no_emails' | ui }}` are not guaranteed to get re-evaluated
when the page language changes. This means that we have to get rid of
the ui filter because it doesn't play well with translations. After this
change-list, we only keep a minimal set of entries in the ui filter to
let us translate the names of the routes in the app.