Explicitly specify a network interface for sync
The Opwen hosts a number of network interfaces, e.g. a wlan interface
to create a local WiFi network to that clients can access the web-app,
the eth interface that is built into the Raspberry Pi by default and an
interface for a cellular modem. Only one of these (namely, the modem)
will actually give us access to the real internet. This means that we
need a way to specify which of the interfaces the upload to Azure Blob
Storage should use so that we don't end up sending our precious data
into nothingness...

NB: The code in the networking module is adapted from a StackOverflow
post: http://stackoverflow.com/a/24196955