Store delivery information in table storage
To make this change concrete, assume that we have three emails:
- Email 1 is sent to "id1@vancouver.ascoderu.ca"
- Email 2 is sent to "id2@victoria.ascoderu.ca"
- Email 3 is sent to "id3@vancouver.ascoderu.ca"

Then we'd store the following in the "domainXdelivered" table:
- `vancouver.ascoderu.ca_False` -> `id1`
- `vancouver.ascoderu.ca_False` -> `id3`
- ` victoria.ascoderu.ca_False` -> `id2`

This gives us fast queries to find all the emails that still have to be
delivered to the `vancouver` client as we can just do a partition scan
on `vancouver.ascoderu.ca_False`. After delivery, we can delete the
entities or flip their delivery bit.