Split storage accounts by service
Throughput units against Azure Blob Storage are shared across all
services in a storage account. This means that if one part of the
application is heavily writing blobs, another part of the application
that's polling queues may see degraded performance.

Splitting the services into separate accounts is a cheap way to get some
more predictable scaling behavior as far as throughput is concerned.