Add instructions for end-to-end setup
We need a few Azure resources to run the email server so we now have a
script in the README that explains how to set them up. We then store the
keys for the resources in a `.env` file and ensure that they can never
be checked-in by git-ignoring the file.

Additionally, we also need at least one registered client so that we can
query the API (e.g. via curl or the API test console), so the script
also sets one up.

Resolves #7