Dequeue messages more frequently
Previously, queue consumers would always wait after processing a
message. This makes sense if there was nothing to process so that we
don't poll the queue too often. However, if the queue consumer did do
work, it's advantageous to continue working right away to avoid build-up
of a backlog on the queue. After this change, a queue consumer will only
sleep in between processing messages if the last dequeue didn't produce
any work.

Additionally, this change also refactors the queue consumer to make it
more testable and avoid having to write tests that rely on timing and
spinning up new threads.